build the parser standalone with its own ctest suite
Split the test setup into tests/CMakeLists.txt and give the project real
options, so the library, the tool and the tests can each be turned off
independently.

Catch2 is now found before it is fetched. A machine that already has
Catch2 v3 installed uses it; anything else falls back to FetchContent, so
a bare `cmake -S . -B build` still works from a clean checkout.

Turn on -Wall -Wextra -Wpedantic. Nothing external is in the include path,
so there are no third-party headers to fight and the library can simply
stay warning-clean. Fix the one signed/unsigned comparison that exposed.

+++ share/dexpr/src/lexer.cpp
@@ -29,7 +29,7 @@ Lexer::Lexer(std::string input)
 }
 
 void Lexer::read_char() {
-  if (read_position_ >= input_.length()) {
+  if (read_position_ >= static_cast<int>(input_.length())) {
     current_char_ = '\0';
   } else {
     current_char_ = input_.at(read_position_);
@@ -39,7 +39,7 @@ void Lexer::read_char() {
 }
 
 char Lexer::peek_char() const {
-  if (read_position_ >= input_.length()) {
+  if (read_position_ >= static_cast<int>(input_.length())) {
     return '\0';
   } else {
     return input_[read_position_];
